feat: Automation to include installer
Testing further automation of copying the installer to the ISO. If the installer project directory is specified, it is synced along with the build files to the ISO. Installer packages are automatically searched for and added to the local ISO repo. Then based on whether the installer is present in any of the repos or not (either from the given project directory or from any other repos in the pacman conf file from the ISO build directory), installer packages are automatically added or removed from the package list file. This will work whether the installer path is specified or not... No duplication of effort - rsync, no extra building of Calamares, etc. This is in addition to other automation for adding AUR packages to the local repo (download PKGBUILD, build the packages from source, and add to the local repo, and add any provided custom pkg.tar.zst file)